split cluster pkg
